+++ R/regexplain-package.R
@@ -1,4 +1,4 @@
 #' @importFrom dplyr "%>%" mutate filter group_by summarize select
-#' @importFrom utils getFromNamespace
+#' @importFrom utils getFromNamespace installed.packages packageVersion
 #' @keywords internal
 "_PACKAGE"